+++ docs/files/ros-action-client.py
@@ -1,0 +1,20 @@
+from __future__ import print_function
+import roslibpy
+import roslibpy.actionlib
+
+client = roslibpy.Ros(host='localhost', port=9090)
+client.run()
+
+action_client = roslibpy.actionlib.ActionClient(client,
+                                                '/fibonacci',
+                                                'actionlib_tutorials/FibonacciAction')
+
+goal = roslibpy.actionlib.Goal(action_client,
+                               roslibpy.Message({'order': 8}))
+
+goal.on('feedback', lambda f: print(f['sequence']))
+goal.send()
+result = goal.wait(10)
+action_client.dispose()
+
+print('Result: {}'.format(result['sequence']))