+++ src/roslibpy/actionlib.py
@@ -10,15 +10,20 @@ can be added. Each goal emits events that can be listened to in order to react t
 updates from the action server. There are four events emmitted: **status**, **result**,
 **feedback**, and **timeout**.
 
+.. autoclass:: Goal
+   :members:
 .. autoclass:: ActionClient
    :members:
-.. autoclass:: Goal
+.. autoclass:: SimpleActionServer
+   :members:
+.. autoclass:: GoalStatus
    :members:
 
 """
 from __future__ import print_function
 
 import logging
+import math
 import random
 import time
 import threading
@@ -27,9 +32,37 @@ from . import Message
 from . import Topic
 from .event_emitter import EventEmitterMixin
 
-__all__ = ['Goal', 'ActionClient']
+__all__ = ['Goal', 'GoalStatus', 'ActionClient', 'SimpleActionServer']
+
+
 LOGGER = logging.getLogger('roslibpy')
 DEFAULT_CONNECTION_TIMEOUT = 3  # in seconds
+
+
+def _is_earlier(t1, t2):
+    """Compares two timestamps."""
+    if t1['secs'] > t2['secs']:
+        return False
+    elif t1['secs'] < t2['secs']:
+        return True
+    elif t1['nsecs'] < t2['nsecs']:
+        return True
+
+    return False
+
+
+class GoalStatus:
+    """Valid goal statuses."""
+    PENDING = 0
+    ACTIVE = 1
+    PREEMPTED = 2
+    SUCCEEDED = 3
+    ABORTED = 4
+    REJECTED = 5
+    PREEMPTING = 6
+    RECALLING = 7
+    RECALLED = 8
+    LOST = 9
 
 
 class Goal(EventEmitterMixin):
@@ -55,6 +88,7 @@ class Goal(EventEmitterMixin):
         self.goal_message = goal_message
         self.goal_id = 'goal_%s_%d' % (random.random(), time.time() * 1000)
 
+        self.wait_result = threading.Event()
         self.result = None
         self.status = None
         self.feedback = None
@@ -94,6 +128,22 @@ class Goal(EventEmitterMixin):
         """Cancel the current goal."""
         self.action_client.cancel_topic.publish(Message({'id': self.goal_id}))
 
+    def wait(self, timeout=None):
+        """Block until the result is available.
+
+        If ``timeout`` is ``None``, it will wait indefinitely.
+
+        Args:
+            timeout (:obj:`int`): Timeout to wait for the result expressed in seconds.
+
+        Returns:
+            Result of the goal.
+        """
+        if not self.wait_result.wait(timeout):
+            raise Exception('Goal failed to receive result')
+
+        return self.result
+
     def _trigger_timeout(self):
         if not self.is_finished:
             self.emit('timeout')
@@ -103,6 +153,7 @@ class Goal(EventEmitterMixin):
 
     def _set_result(self, result):
         self.result = result
+        self.wait_result.set()
 
     def _set_feedback(self, feedback):
         self.feedback = feedback
@@ -220,35 +271,227 @@ class ActionClient(EventEmitterMixin):
         if not self.omit_result:
             self.result_listener.unsubscribe()
 
+class SimpleActionServer(EventEmitterMixin):
+    """Implementation of the simple action server.
 
-if __name__ == '__main__':
-    import logging
-    from . import Ros
+    The server emits the following events:
 
-    FORMAT = '%(asctime)-15s [%(levelname)s] %(message)s'
-    logging.basicConfig(level=logging.DEBUG, format=FORMAT)
+    * ``goal``: fires when a new goal has been received by the server.
+    * ``cancel``: fires when the client has requested the cancellation of the action.
 
-    ros_client = Ros('127.0.0.1', 9090)
+    Args:
+        ros (:class:`.Ros`): Instance of the ROS connection.
+        server_name (:obj:`str`): Action server name, e.g. ``/fibonacci``.
+        action_name (:obj:`str`): Action message name, e.g. ``actionlib_tutorials/FibonacciAction``.
+    """
+    STATUS_PUBLISH_INTERVAL = 0.5   # In seconds
 
-    def handle_result(result, client):
-        print('Final result', result['sequence'])
+    def __init__(self, ros, server_name, action_name):
+        super(SimpleActionServer, self).__init__()
 
-        client.dispose()
-        client.ros.call_later(2, client.ros.close)
+        self.ros = ros
+        self.server_name = server_name
+        self.action_name = action_name
+        self._lock = threading.Lock()
 
-    def run_action_example():
-        action_client = ActionClient(ros_client, '/fibonacci', 'actionlib_tutorials/FibonacciAction')
-        goal = Goal(action_client, Message({'order': 6}))
+        # Create all required publishers and listeners
+        self.feedback_publisher = Topic(ros, server_name + '/feedback', action_name + 'Feedback')
+        self.status_publisher = Topic(ros, server_name + '/status', 'actionlib_msgs/GoalStatusArray')
+        self.result_publisher = Topic(ros, server_name + '/result', action_name + 'Result')
+        self.goal_listener = Topic(ros, server_name + '/goal', action_name + 'Goal')
+        self.cancel_listener = Topic(ros, server_name + '/cancel', 'actionlib_msgs/GoalID')
 
-        goal.on('result', lambda result: handle_result(result, action_client))
-        goal.on('feedback', lambda feedback: print(feedback))
-        goal.on('timeout', lambda: print('TIMEOUT'))
+        # Advertise all publishers
+        self.feedback_publisher.advertise()
+        self.status_publisher.advertise()
+        self.result_publisher.advertise()
 
-        goal.send(60)
+        # Track the goals and their status in order to publish status
+        self.status_message = Message(dict(
+            header=dict(
+                stamp=dict(secs=0, nsecs=100),
+                frame_id=''
+            ),
+            status_list=[]
+        ))
 
-    ros_client.on_ready(run_action_example, run_in_thread=True)
+        # Needed for handling preemption prompted by a new goal being received
+        self.current_goal = None    # currently tracked goal
+        self.next_goal = None       # the one that'll be preempting
+        self.preempt_request = False
 
-    try:
-        ros_client.run_forever()
-    except KeyboardInterrupt:
-        ros_client.terminate()
+        self.goal_listener.subscribe(self._on_goal_message)
+        self.cancel_listener.subscribe(self._on_cancel_message)
+
+        # Intentionally not publishing immediately and instead
+        # waiting one interval for the first message
+        self.ros.call_later(self.STATUS_PUBLISH_INTERVAL, self._publish_status)
+
+    def start(self, action_callback):
+        """Start the action server.
+
+        Args:
+            action_callback: Callable function to be invoked when a new goal is received. It takes one paramter containing the goal message.
+        """
+        LOGGER.info('Action server {} started'.format(self.server_name))
+
+        def _internal_goal_callback(goal):
+            LOGGER.info('Action server {} received new goal'.format(self.server_name))
+            self.ros.call_in_thread(lambda: action_callback(goal))
+
+        def _internal_preempt_callback():
+            LOGGER.info('Action server {} received preemption request'.format(self.server_name))
+            self.preempt_request = True
+
+        self.on('goal', _internal_goal_callback)
+        self.on('cancel', _internal_preempt_callback)
+
+    def _publish_status(self):
+        # Status publishing is required for clients to know they've connected
+        with self._lock:
+            current_time = time.time()
+            secs = int(math.floor(current_time))
+            nsecs = int(round(1e9 * (current_time - secs)))
+
+            self.status_message['header']['stamp']['secs'] = secs
+            self.status_message['header']['stamp']['nsecs'] = nsecs
+            self.status_publisher.publish(self.status_message)
+
+        # Invoke again in the defined interval
+        self.ros.call_later(self.STATUS_PUBLISH_INTERVAL, self._publish_status)
+
+    def _on_goal_message(self, message):
+        will_cancel = False
+        will_emit_goal = None
+
+        with self._lock:
+            if self.current_goal:
+                self.next_goal = message
+                # needs to happen AFTER rest is set up
+                will_cancel = True
+            else:
+                self.status_message['status_list'] = [
+                    dict(goal_id=message['goal_id'], status=GoalStatus.ACTIVE)]
+                self.current_goal = message
+                will_emit_goal = message['goal']
+
+        if will_cancel:
+            self.emit('cancel')
+        if will_emit_goal:
+            self.emit('goal', will_emit_goal)
+
+    def _on_cancel_message(self, message):
+        # As described in the comments of the original roslibjs code
+        # this may be more complicated than necessary, since it's
+        # not sure the callbacks can ever wind up with a scenario
+        # where we've been preempted by a next goal, it hasn't finished
+        # processing, and then we get a cancel message.
+
+        will_cancel = False
+
+        with self._lock:
+            message_id = message['id']
+            message_stamp = message['stamp']
+            secs = message['stamp']['secs']
+
+            if secs == 0 and secs == 0 and message_id == '':
+                self.next_goal = None
+                if self.current_goal:
+                    will_cancel = True
+
+            else: # treat id and stamp independently
+                if self.current_goal and message_id == self.current_goal['goal_id']['id']:
+                    will_cancel = True
+                elif self.next_goal and message_id == self.next_goal['goal_id']['id']:
+                    self.next_goal = None
+
+                if self.next_goal and _is_earlier(self.next_goal['goal_id']['stamp'], message_stamp):
+                    self.next_goal = None
+                if self.current_goal and _is_earlier(self.current_goal['goal_id']['stamp'], message_stamp):
+                    will_cancel = True
+
+        if will_cancel:
+            self.emit('cancel')
+
+    def is_preempt_requested(self):
+        """Indicate whether the client has requested preemption of the current goal."""
+        with self._lock:
+            return self.preempt_request
+
+    def set_succeeded(self, result):
+        """Set the current action state to succeeded.
+
+        Args:
+            result (:obj:`dict`): Dictionary of key/values to set as the result of the action.
+        """
+        LOGGER.info(
+            'Action server {} setting current goal to SUCCEEDED'.format(self.server_name))
+
+        with self._lock:
+            result_message = Message({
+                'status': {
+                    'goal_id': self.current_goal['goal_id'],
+                    'status': GoalStatus.SUCCEEDED
+                },
+                'result': result
+            })
+
+            self.result_publisher.publish(result_message)
+
+            self.status_message['status_list'] = []
+
+            if self.next_goal:
+                self.current_goal = self.next_goal
+                self.next_goal = None
+            else:
+                self.current_goal = None
+
+            self.preempt_request = False
+
+        # If there's a new current goal assigned, emit it
+        if self.current_goal:
+            self.emit('goal', self.current_goal['goal'])
+
+    def send_feedback(self, feedback):
+        """Send feedback.
+
+        Args:
+            feedback (:obj:`dict`): Dictionary of key/values of the feedback message.
+        """
+        feedback_message = Message({
+            'status': {
+                'goal_id': self.current_goal['goal_id'],
+                'status': GoalStatus.ACTIVE
+            },
+            'feedback': feedback
+        })
+
+        self.feedback_publisher.publish(feedback_message)
+
+    def set_preempted(self):
+        """Set the current action to preempted (cancelled)."""
+        LOGGER.info('Action server {} preempting current goal'.format(self.server_name))
+
+        with self._lock:
+            self.status_message['status_list'] = []
+            result_message = Message({
+                'status': {
+                    'goal_id': self.current_goal['goal_id'],
+                    'status': GoalStatus.PREEMPTED
+                }
+            })
+
+            self.result_publisher.publish(result_message)
+
+            if self.next_goal:
+                self.current_goal = self.next_goal
+                self.next_goal = None
+            else:
+                self.current_goal = None
+
+            # Preemption completed
+            self.preempt_request = False
+
+        # If there's a new current goal assigned, emit it
+        if self.current_goal:
+            self.emit('goal', self.current_goal['goal'])